+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Blip is the web front end for YourLoops system.
 It is based on Tidepool Blip 1.27.
 
 ## 3.7.0-beta - 2025-03-05
+### Added
+- YLP-3151 Display Night mode events in Daily view
+
 ### Changed
 - YLP-3478 Update the front of the maintenance page
 - YLP-3486 Add physical activity name in tooltip